backend/pyproject.toml
[project]
name = "pingsight-backend"
version = "0.1.0"
description = "FastAPI backend for pingSight"
requires-python = ">=3.10"

dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.1",
  "sqlalchemy>=2.0.30",
  "alembic>=1.13.2",
  "psycopg[binary]>=3.2.1",
  "httpx>=0.27.0",
  "pydantic-settings>=2.3.4",
  "python-multipart>=0.0.9",
  "python-jose[cryptography]>=3.3.0",
  "pyjwt>=2.8.0",
  "passlib[bcrypt]>=1.7.4",
  "dotenv>=0.9.9",
  "fastapi-sso>=0.21.0",
  "python-dotenv>=1.2.2",
  "apscheduler>=3.10.4",
]

[tool.uv]
package = false
